Solve 2021 day 21
The trick here was to realise that there aren't that many possible total
different states - only 10 x 10 x 22 x 22, and that once two games fall into the
same state, their future will be the same, even if they came from different
places. Tracking the number of games in each state as the universes evolve is
therefore the way to go.