Formatting fix and small refactor
I ran black on all the code blocks, so the formatting is consistent. I also
updated the `load` function in the `utils` file so that by default it strips
newlines from the strings it returns; that meant that the stripping logic could
be moved out of all the individual files.